refactor: introduce renderer class
