+++ package.json
@@ -27,8 +27,7 @@
   "type": "module",
   "dependencies": {
     "axios": "^1.10.0",
-    "commander": "^11.1.0",
-    "picocolors": "^1.1.1"
+    "commander": "^11.1.0"
   },
   "bin": {
     "spectest": "dist/cli.js"
